refactor: extract protocol translation seam
